+++ src/main/java/net/spell_engine/api/spell/event/SpellHandlers.java
@@ -20,7 +20,7 @@ public class SpellHandlers {
     public static final Map<String, CustomDelivery> customDelivery = new HashMap<>();
     public interface CustomDelivery {
         boolean onSpellDelivery(World world, RegistryEntry<Spell> spellEntry, PlayerEntity caster,
-                                List<SpellHelper.TargetWithContext> targets, SpellHelper.ImpactContext context,
+                                List<SpellHelper.DeliveryTarget> targets, SpellHelper.ImpactContext context,
                                 @Nullable Vec3d targetLocation);
     }
     public static void registerCustomDelivery(Identifier id, CustomDelivery delivery) {